feat: remove useless code and tidy up code layout for dbctl
* playground only supports wesql engine, remove mysql operator

* Importing external code to implement the bench instead of copying

* Code related to playground is placed in a directory
