+++ apis/dataprotection/v1alpha1/backupjob_types.go
@@ -115,6 +115,7 @@ type BackupJobStatus struct {
 // +kubebuilder:printcolumn:name="Phase",type=string,JSONPath=`.status.phase`
 // +kubebuilder:printcolumn:name="CompletionTime",type=date,JSONPath=`.status.completionTimestamp`
 // +kubebuilder:printcolumn:name="Age",type=date,JSONPath=`.metadata.creationTimestamp`
+
 // BackupJob is the Schema for the backupjobs API (defined by User)
 type BackupJob struct {
 	metav1.TypeMeta   `json:",inline"`
