+++ apis/dataprotection/v1alpha1/backuptool_types.go
@@ -87,6 +87,7 @@ type BackupToolStatus struct {
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
 // +kubebuilder:resource:categories={dbaas},scope=Cluster
+
 // BackupTool is the Schema for the backuptools API (defined by ISV)
 type BackupTool struct {
 	metav1.TypeMeta   `json:",inline"`
