+++ apis/dataprotection/v1alpha1/restorejob_types.go
@@ -87,6 +87,7 @@ type RestoreJobStatus struct {
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
 // +kubebuilder:resource:categories={dbaas},scope=Namespaced
+
 // RestoreJob is the Schema for the restorejobs API (defined by User)
 type RestoreJob struct {
 	metav1.TypeMeta   `json:",inline"`
