+++ controllers/dataprotection/restorejob_controller.go
@@ -297,7 +297,6 @@ func (r *RestoreJobReconciler) GetPodSpec(reqCtx intctrlutil.RequestCtx, restore
 	return podSpec, nil
 }
 
-// GetPodSpec(reqCtx intctrlutil.RequestCtx, restoreJob *dataprotectionv1alpha1.RestoreJob) (corev1.PodSpec, error) {
 func (r *RestoreJobReconciler) PatchTargetCluster(reqCtx intctrlutil.RequestCtx, restoreJob *dataprotectionv1alpha1.RestoreJob, patch []byte) error {
 	// get stateful service
 	clusterTarget := &appv1.StatefulSetList{}
