+++ pkg/cmd/bench/tpcc.go
@@ -26,10 +26,10 @@ import (
 
 	"github.com/pingcap/go-tpc/pkg/measurement"
 	"github.com/pingcap/go-tpc/pkg/workload"
+	"github.com/pingcap/go-tpc/tpcc"
 	"github.com/spf13/cobra"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
 
-	"github.com/apecloud/kubeblocks/pkg/cmd/bench/tpcc"
 	"github.com/apecloud/kubeblocks/pkg/utils"
 )
