+++ pkg/cmd/playground/wesql.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package provider
+package playground
 
 import (
 	"helm.sh/helm/v3/pkg/repo"
