+++ pkg/utils/print.go
@@ -20,61 +20,8 @@ import (
 	"os"
 	"text/template"
 
-	"github.com/fatih/color"
-
 	"github.com/apecloud/kubeblocks/pkg/types"
 )
-
-var (
-	red            = color.New(color.FgRed).SprintFunc()
-	green          = color.New(color.FgGreen).SprintFunc()
-	k3dImageStatus = map[string]bool{}
-	x              = red("✘")
-	y              = green("✔")
-)
-
-type PlayGroundInfo struct {
-	DBCluster     string
-	DBPort        string
-	DBNamespace   string
-	Namespace     string
-	ClusterName   string
-	GrafanaSvc    string
-	GrafanaPort   string
-	GrafanaUser   string
-	GrafanaPasswd string
-	HostIP        string
-	CloudProvider string
-	Version       string
-}
-
-type DBClusterInfo struct {
-	DBCluster       string
-	DBPort          string
-	Version         string
-	Topology        string
-	Status          string
-	StartTime       string
-	Labels          string
-	RootUser        string
-	DBNamespace     string
-	Instances       int64
-	ServerId        int64
-	Secret          string
-	OnlineInstances int64
-	Storage         int64
-	Engine          string
-	HostIP          string
-}
-
-type BackupJobInfo struct {
-	Name           string
-	Namespace      string
-	Phase          string
-	StartTime      string
-	CompletionTime string
-	Labels         string
-}
 
 var playgroundTmpl = `
 Notes:
@@ -144,59 +91,11 @@ Connect:        mysql -h {{.HostIP}} -u{{.RootUser}} -p"$MYSQL_ROOT_PASSWORD"
 
 `
 
-func PrintClusterStatus(status types.ClusterStatus) bool {
-	InfoP(0, "K3d images status:")
-	if status.K3dImages.Reason != "" {
-		Info(x, "K3d images:", status.K3dImages.Reason)
-		return true // k3d images not ready
-	}
-	k3dImageStatus[types.K3sImage] = status.K3dImages.K3s
-	k3dImageStatus[types.K3dToolsImage] = status.K3dImages.K3dTools
-	k3dImageStatus[types.K3dProxyImage] = status.K3dImages.K3dProxy
-	stop := false
-	for i, imageStatus := range k3dImageStatus {
-		stop = stop || !imageStatus
-		if !imageStatus {
-			InfoP(1, x, "image", i, "not ready")
-		} else {
-			InfoP(1, y, "image", i, "ready")
-		}
-	}
-	if stop {
-		return stop
-	}
-	InfoP(0, "Cluster(K3d) status:")
-	if status.K3d.Reason != "" {
-		Info(x, "K3d:", status.K3d.Reason)
-		return true // k3d not ready
-	}
-	for _, c := range status.K3d.K3dCluster {
-		cr := x
-		if c.Reason != "" {
-			InfoP(1, x, "cluster", "[", c.Name, "]", "not ready:", c.Reason)
-			stop = true
-		} else {
-			InfoP(1, y, "cluster", "[", c.Name, "]", "ready")
-			cr = y
-		}
-
-		// Print helm release status
-		for k, v := range c.ReleaseStatus {
-			InfoP(2, cr, "helm chart [", k, "] status:", v)
-		}
-	}
-	if stop {
-		return stop
-	}
-
-	return false
-}
-
-func PrintPlaygroundGuild(info PlayGroundInfo) error {
+func PrintPlaygroundGuide(info types.PlaygroundInfo) error {
 	return PrintTemplate(playgroundTmpl, info)
 }
 
-func PrintClusterInfo(info *DBClusterInfo) error {
+func PrintClusterInfo(info *types.DBClusterInfo) error {
 	return PrintTemplate(clusterInfoTmpl, info)
 }
